refactor: Moved interface and email validator to Domain
